Realizado atualizações para que nao seja permitido deletar ou editar usuário master

+++ src/main/java/com/dougfsilva/iotizzy/service/user/DeleteUser.java
@@ -1,8 +1,10 @@
 package com.dougfsilva.iotizzy.service.user;
 
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
+import com.dougfsilva.iotizzy.exception.OperationNotPermittedException;
 import com.dougfsilva.iotizzy.model.User;
 import com.dougfsilva.iotizzy.mqtt.commands.DeleteClientMqttCommand;
 import com.dougfsilva.iotizzy.repository.UserRepository;
@@ -12,6 +14,9 @@ import com.dougfsilva.iotizzy.service.measuringDevice.DeleteMeasuringDevice;
 @Service
 @PreAuthorize("hasAnyRole('GOLD_USER', 'SILVER_USER', 'ADMIN')")
 public class DeleteUser {
+	
+	@Value("${user.master.email}")
+	private String masterEmail;
 	
     private final UserRepository repository;
     private final DeleteClientMqttCommand deleteClientMqtt;
@@ -31,6 +36,9 @@ public class DeleteUser {
 
 	public void delete(){
         User user = authenticatedUser.getUser();
+        if(user.getEmail().getAddress().equals(masterEmail)) {
+        	throw new OperationNotPermittedException("It is not allowed to delete the master user");
+        }
         repository.delete(user);
         deleteClientMqtt.delete(user);
         deleteMeasuringDevice.deleteAll();
